start work on generateCode and pipeEx
Start defining generateCode. Although, temporarily comment it out
and work from the vertex upwards with generateCodeFromVertex.

Remove the free variable from the StreamOperator definition. It
was superfluous now that we have the parameters member for
StreamVertex.

comment out the test_* definitions which broke as a result of the
above; to be re-instated once I've understood why they have broken.

Add a pipeEx stream graph definition. This roughly corresponds to
"s4" in CompileIoT, as well as the static example in examples/pipeline.
The hope here is I can iteratively adjust the definition of pipeEx,
generateCode etc. until I can output something which is equivalent
to the static example.