(temporarily) add vertex id to show implementation
Without this, due to a bug/bad assumptions elsewhere, the graphviz
export of a streamgraph would have some nodes merged incorrectly.
E.g. in the Merge example, the two vertices that are "streamMap(id)"
would be collapsed.

The correct solution would be to not conflate node IDs and labels
in the graphviz export; but for now, make the labels unique by
prefixing their unique IDs.