generateCodeFromStreamGraph: more use of valence
Adjust partValence to consider the number of Source-type nodes as
well as edges into the graph in question: So a graph with no edges
into it from other graphs, but one or more Source-type nodes will
have a valence of one (or more)

When constructing streamGraphFn, consider valence to determine how
many arguments the function should accept. Correct the calculation
of lastIdentifier accordingly.

Add a note that the generation of code for (at least) Merge-type
vertices is not correct and we will need to consider each operator
separately.