operator-specific logic to generateCodeFromVertex
in the case of the Merge operator, do not calculate and append
stream graph arguments. Assume that the parameter value covers
them all.

There is some tension/ambiguity regarding paremeters, as supplied
in the streamgraph definition, and the calculated parameters for
streamgraph inter-linking. Perhaps a better approach would be to
always calculate the inter-linking parameters, and ignore the
"parameters" field for Merge operations. There's a similar issue
with the way Merge expects its arguments (one argument, a list to
merge) and the way it will practically be wired up (via nodeSink2
or similar, where the arguments are separated out).