Fix mergeMap testing
Typos in the test (copying from the previous test) led me to think
that ordering was not being preserved! It is, but I've had to use
an alternative unary Char function to succ (from Enum) which is not
valid for all Char values. I went with "isAscii" which doesn't
generate very interesting output (streams of Bools) but at least
works.

The test is still expensive to evaluate so I've left the property
defined as pxxp_* so HTF doesn't collect it.