fix promotion of streamFilter → streamFilterAcco
the "promoted" function was defined in such a way that the type of the
stream was unified with the type of the dummy accumulator, which I was
forcing to Char. Thus it could only be used for Streams of Chars.

Fix this and pull out the predicate so it is more generic. Uncomment
a quickCheck property to prove that the type of the accumulator was
unimportant.