demonstrate drawback of pattern matching; workaround
Demonstrate that pattern matching only works if the graph you are
matching against is constructed in exactly the same fashion, but
they may not be.

Demonstrate that Algebra.Graph.simplify addresses this for the two
test cases. But will it in all circumstances?

Bring in HTF to aid in defining and running the tests