attempt recursive application of rewrite rule
Unfortunately this uncovers a shortcoming of this approach: even
when a rewrite-pattern matches, and the rewrite is performed, there
may be other rewriting that needs to take place in the graph.
Consider

    filter p1 → map f → filter p2

encoded as

    g = filter p1 * map f + map f * filter p2

applying the mapFilter rewrite will match the second half, and rewrite
it, but not touch map f in the first half

    filter p1 * map f + filter (p2.f) * map f

corresponding to

    filter p1 → map f ← filter (p2.f)

So when we detect a match for a rewrite rule in a sub-graph, we are
going to need to perform rewriting at the top-level.