Convert writeup.adoc to LHS (TeX)
and slowly start migrating code out of Rewrites.hs into it

I'm not sure whether this is a good use of time right now; battling
TeX feels like a distraction.

Another approach might be to inline code in the AsciiDoc and see if
its possible to get AsciiDoc to ignore the \begin{code} and \end{code}
blocks for its output.