convert writeup.adoc to Literate Haskell (.lhs)
Inline haskell code is wrapped in \begin{code} and \end{code}, and
GHC ignores everything else.

Code that we do not want to include in the output Document can be
wrapped in AsciiDoc comment lines: /////…

Code that we do want to include can be marked up as a code block by
wrapping it in dash-lines: ----…

Two hacks minorly improve the output:

 • by moving the dash-lines *inside* the \begin{code}/\end{code}
   delimiters, the literate-haskell delimiters do not show up in
   the Document's code block. The dash-lines are interpreted as
   comments by GHC, and ignored.

 • The begin/end{code} delimiters remain in the Document output,
   although they are better outside of the code blocks. They can
   be partially disguised by defining an empty AsciiDoc macro for
   {code}: that part is then substituted, leaving just stray
   \begin and \end.

On the whole this is still a bit faster to work with than a conversion
to LaTeX. At least for me, right now.