displayGraph: use pipes, avoid temp files
use pipes to connect the Haskell process to "dot" and "dot"
to "display"; then write the converted stream graph string
to the pipeline. Avoid any temporary files.

I think this is using general enough interfaces that it
should work on Windows, so long as "dot" and "display" are
in the $PATH (or %PATH%)