StreamVertexQ type, simpleStream adjustment
filter example as our guinea-pig. simpleStream gives a Graph
(StreamVertexQ) which we can "deQ" into a StreamGraph in IO ().
Do this in the main method and pass the result on to partitionGraph
as usual. (later work: adjust partitionGraph).

deQ uses foldG to translate the ExpQ -> Exp. The exact order of
vertices is unknown/undefined/??, but at least there's a consistent
environment that they are all evaluated within.