rework StreamVertex params as ExpQ
For CompileIoT, cheat a little bit and use unsafePerformIO to unwrap
the Exp. This is a short-term fix (which works!), as CompileIoT will
benefit from composing a large Exp(Q) rather than mashing strings
together as it does now.

Adjust examples/filter accordingly, the generated files are valid
Haskell and the whole example works end-to-end.

Testing rewrites is impaired a little bit by not being able to compare
the parameters. We can probably test other properties of them.