CompileIoT: derive requiredModules from parameters
Via the "Scrap your boilerplate" (syb) generic programming package,
derive a list of all Names referenced in the parameter Expressions
for the StreamGraph. Use this to add to the list of modules to
import in generated code.

This saves the programmer from having to write a long import list
themselves. Shorten the explicit import list for the Taxi example.