Adjust Jackson stuff into Taxi example
Move some of the Taxi example Jackson constants into Taxi/generate.hs.
Adjust module export lists accordingly.

Tweak the fields of JacksonParams structure to only include the three
values we cannot infer.

Write calcInputs to calculate the "Inputs" array from a StreamGraph,
and a basic test.

We are now probably in a position to work passing these things through
the normal call stack and write a cost model based on the Jackson data.