fix calcAllSg
I had mixed up the "inputs" array (external arrivals and proportion of
global arrival rates) with the calculated "arrivals" array.

Tweak the taxi example numbers so they match up between the two
different definitions (one in Taxi, one in Jackson). Export the Jackson
definition so we can import it and compare it in the Taxi example.

Note that the calculation of some of the matrices is heavily reliant on
the stream graph being very simple (one source node at the start): I
will need to return to this and fix it up.

Write some HTF tests to demonstrate that the streamgraph-based entry
points to Jackson are definitely equivalent to the originals.