rework derive* functions to better handle sources
Instead of trying to chop off the source node(s), and guessing there's
only one, leave them in. Build appropriately-bounded Arrays for the
range defined by the supplied vertex Ids. The Arrays cannot be sparse,
so populate the invalid pairings with 0 values (or the equivalent).

This results in breakages elsewhere: starting with arrivalRate
calculation. Next job is fix that.