WIP: Initial separate builder/runner images
Define a new container, striot/striot-runtime, intended to be used as
the basis for runtime containers. Based on Debian 10 slim (aka Buster,
the current stable version at the time of writing) with the additional
librdkafka1 installed.

Adjust the Dockerfile-generation code to use a multi-stage build, where
the build takes place in striot/striot-base, and the resulting "node"
binary copied over to a new container based on the runtime.

The resulting images are much smaller: striot/striot-base is 3.15GB
right now, and striot/striot-runtime would be about 93.5MB. The binaries
we build are typically around 50M, but could be reduced to about 33M if
we strip the debug symbols out.

striot-base is based on haskell:8.6, which is based on Debian "Stretch"
("oldstable", the release before the current stable). There could be ABI
incompatibilities between libc or librdkafka1 in the builder and runner
containers. We should harmonize on a common base version; possibly the
/next/ Debian stable release (bullseye), which is due very soon, once
there are hub.docker.com/_/haskell containers based on it. I've marked
this "WIP" for that reason. (We might also need to look at GHC version.
I think 8.6 is already "unsupported" in the containers)

Another reason this is WIP: it will break the Taxi example, since node1
requires "sorteddata.csv" in the container.
debian:stretch