WIP: adjust mapWindow gating for polymorphic types
Relax the restriction on types (requiring the map to have the same
input and output type) to allow for the situation where the window
type is polymorphic, and so could be applied to the map input type
when re-ordered.

Add some basic utility functions for type reasoning that operate
on the types-encoded-as-strings.

Unfinished test compatTtest

Expose filterAccWindow

Rewrite the test graph in WearableExample to avoid "simpleStream",
since that infers the input type (and would infer a concrete value,
preventing mapWindow from applying)

We can now reproduce the rewriting that Path2IOT did.