WIP: adjust optimiser to maintain tree of rewrites
Adjust Logical Optmiser to maintain a record of which rewrite rules
produced a variant StreamGraph.

Introduce a LabelledRewriteRule type and update our lists of rewrite
rules to be labelled.

Introduce a Variant (|Original) type which groups together

 * a StreamGraph
 * a Label for the rewrite Rule that produced it
 * a parent Variant

Accessor variantGraph' that handles Variant|Original (awkward name
since I used variantGraph in the accessor for Variant)

New function applyRules2, a variation of applyRules which operates
on Variants instead of StreamGraphs.